Flip the dependency between editor and theme
