+++ crates/editor/Cargo.toml
@@ -11,6 +11,7 @@ buffer = { path = "../buffer" }
 clock = { path = "../clock" }
 gpui = { path = "../gpui" }
 sum_tree = { path = "../sum_tree" }
+theme = { path = "../theme" }
 util = { path = "../util" }
 anyhow = "1.0"
 lazy_static = "1.4"
